Maybe simplification was inaccurate. But this should give us a bit of a speed boost. This does not work but this sets up the basic code to debug.

+++ js/a-sky-forge/src/glsl/moon_vertex.glsl
@@ -1,6 +1,6 @@
 #ifdef GL_ES
-precision highp float;
-precision highp int;
+precision mediump float;
+precision mediump int;
 #endif
 
 varying vec3 vWorldPosition;
